appDir から pages に変更
