+++ next.config.js
@@ -1,8 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  experimental: {
-    appDir: true,
-  },
+  output: 'export',
+  reactStrictMode: true,
 }
 
 module.exports = nextConfig
